core: rename folder 'release' to 'models', #TASK-591
